+++ packages/args/src/index.ts
@@ -6,11 +6,13 @@
 import format from './format';
 import parse from './parse';
 import parseInContext from './parseInContext';
-import ArgsError, { ArgsErrorCode } from './ArgsError';
+import ArgsError from './ArgsError';
+import type { ArgsErrorCode } from './ArgsError';
 import ParseError from './ParseError';
 import ValidationError from './ValidationError';
 
 export * from './constants';
 export * from './types';
 
-export { format, parse, parseInContext, ArgsError, ArgsErrorCode, ParseError, ValidationError };
+export { format, parse, parseInContext, ArgsError, ParseError, ValidationError };
+export type { ArgsErrorCode };